Implement component for wrap the lyric row table logic. This class will handler the selecting state and the display the add new lyric area.
